refactor: moved db.ts to server folder
